backend/requirements.txt
asgiref==3.11.1
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.61.1
pillow==12.1.1
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.0.1
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2025.3
xhtml2pdf>=0.2.11
webencodings==0.5.1
zopfli==0.4.1
